move types
- move s_sound_globals_definition and s_game_globals_multiplayer_information out of game_globals
- remove header includes in particle files
- move ui error types into their own header
