+++ xlive/Blam/Engine/interface/screens/screen_4way_signin.cpp
@@ -14,6 +14,7 @@
 #include "interface/user_interface.h"
 #include "interface/user_interface_bitmap_block.h"
 #include "interface/user_interface_controller.h"
+#include "interface/user_interface_errors.h"
 #include "interface/user_interface_memory.h"
 #include "interface/user_interface_model_block.h"
 #include "interface/user_interface_networking.h"
