+++ xlive/Blam/Engine/interface/screens/screen_display_mode.cpp
@@ -2,8 +2,9 @@
 #include "screen_display_mode.h"
 #include "screen_confirm_resolution.h"
 
-#include "interface/user_interface_memory.h"
 #include "interface/user_interface_controller.h"
+#include "interface/user_interface_errors.h"
+#include "interface/user_interface_memory.h"
 #include "rasterizer/rasterizer_settings.h"
 
 /* constants */
