+++ xlive/Blam/Engine/interface/screens/screen_squad_settings.cpp
@@ -7,10 +7,11 @@
 
 #include "bitmaps/bitmap_group.h"
 #include "cache/cache_files.h"
-#include "interface/user_interface_controller.h"
-#include "interface/user_interface_networking.h"
-#include "interface/user_interface_memory.h"
 #include "interface/user_interface_bitmap_block.h"
+#include "interface/user_interface_controller.h"
+#include "interface/user_interface_errors.h"
+#include "interface/user_interface_memory.h"
+#include "interface/user_interface_networking.h"
 #include "interface/user_interface_screen_widget_definition.h"
 #include "main/levels.h"
 #include "main/level_definitions.h"
