+++ xlive/Project_Cartographer.vcxproj.filters
@@ -2104,6 +2104,8 @@
     <ClInclude Include="Blam\Engine\kablam\kablam_shell\kablam_shell.h" />
     <ClInclude Include="Blam\Engine\simulation\simulation_update.h" />
     <ClInclude Include="Blam\Engine\game\player_training.h" />
+    <ClInclude Include="Blam\Engine\game\multiplayer_definitions.h" />
+    <ClInclude Include="Blam\Engine\interface\user_interface_errors.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="resources\Resource.rc" />
